requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo[srv]==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
gunicorn==21.2.0
slowapi==0.1.9 